requirements.txt
# Core backend
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.7.4
python-dotenv==1.0.0

# Modelos de IA — adaptadores directos
google-genai==1.0.0
openai==1.30.0
anthropic==0.28.0

# LangChain — para el adaptador de contraste pedagógico
langchain-core==0.3.0
langchain-google-genai==2.0.0
